Tweaked resource consumption and fixed swamp hangup
Discovered large swamps can cause pathfinding errors for screeps on room
edge. Fix for creeps with bug coming in from the west side, can be seen
in claim code. Removed code for previous fixes as it was not actually
helping. Set distributers to take from the link if the storage is empty.
So the death of a linker wont cause as many issues.